# [P4RT]Update P4RT test to include on-SUT multicast programming,Port ID assignment and init time prior to performance testing.