--- a/p4rt_app/scripts/BUILD.bazel
+++ b/p4rt_app/scripts/BUILD.bazel
@@ -65,6 +65,7 @@ cc_binary(
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/strings:str_format",
         "@com_google_absl//absl/time",
+        "@com_google_absl//absl/types:span",
         "@com_google_googletest//:gtest_main",
     ],
 )