**Keycheck Result:**

sonic-buildimage/src/sonic-p4rt/sonic-pins$ ~/tools/keyword_check.sh ./p4rt_app/ .
Keyword check Passed.

**Bazel Build Output:**
/sonic/src/sonic-p4rt/sonic-pins$ bazel build $BAZEL_BUILD_OPTS ...
INFO: Analyzed 717 targets (0 packages loaded, 0 targets configured).
INFO: Found 717 targets...
INFO: From Compiling p4rt_app/scripts/p4rt_route_measurement_test.cc:
p4rt_app/scripts/p4rt_route_measurement_test.cc: In function 'absl::lts_20230802::Status p4rt_app::{anonymous}::VerifyP4WriteRequestSizes(const P4WriteRequests&, uint32_t, uint32_t, bool)':
p4rt_app/scripts/p4rt_route_measurement_test.cc:307:21: warning: comparison of integer expressions of different signedness: 'int' and 'std::vector<p4::v1::WriteRequest>::size_type' {aka 'long unsigned int'} [-Wsign-compare]
  307 |   for (int i = 0; i < requests.inserts.size(); ++i) {
      |                   ~~^~~~~~~~~~~~~~~~~~~~~~~~~
p4rt_app/scripts/p4rt_route_measurement_test.cc:309:13: warning: comparison of integer expressions of different signedness: 'int' and 'uint32_t' {aka 'unsigned int'} [-Wsign-compare]
  309 |     if (got != expected_requests_per_batch) {
      |                                            ~~~~~~~~~~~~~~^~~~~~
INFO: Elapsed time: 45.445s, Critical Path: 44.42s
INFO: 6 processes: 2 internal, 4 linux-sandbox.
INFO: Build completed successfully, 6 total actions


**Bazel Test output:**

/sonic/src/sonic-p4rt/sonic-pins$ bazel test $BAZEL_BUILD_OPTS --cache_test_results=no ...
INFO: Analyzed 717 targets (0 packages loaded, 0 targets configured).
INFO: Found 492 targets and 225 test targets...
INFO: Elapsed time: 216.208s, Critical Path: 121.44s
INFO: 282 processes: 339 linux-sandbox, 18 local.
INFO: Build completed successfully, 282 total actions
//dvaas:port_id_map_test                                                 PASSED in 0.8s
//dvaas:test_run_validation_golden_test                                  PASSED in 0.1s
//tests/lib:packet_generator_test                                        PASSED in 64.6s
//p4rt_app/tests:packet_replication_table_test                           PASSED in 3.1s
//p4rt_app/tests:packetio_test                                           PASSED in 4.9s
//p4rt_app/tests:port_name_and_id_test                                   PASSED in 4.9s
  Stats over 4 runs: max = 64.6s, min = 61.9s, avg = 63.1s, dev = 1.0s
//sai_p4/instantiations/google/tests:p4_constraints_integration_test     PASSED in 0.9s
  Stats over 5 runs: max = 0.9s, min = 0.7s, avg = 0.8s, dev = 0.1s
//sai_p4/instantiations/google/test_tools:table_entry_generator_test     PASSED in 46.0s
  Stats over 50 runs: max = 46.0s, min = 0.8s, avg = 1.9s, dev = 6.4s
Executed 225 out of 225 tests: 225 tests pass.
INFO: Build completed successfully, 282 total actions